Create monte carlo.R
Okay, we're getting rid of the version numbers.

This is the basic monte carlo model code plus the ability to have spatial variation with respect to temperature.

We should have a conversation about structuring our computers in the same way so I can upload all of the data files to git too and change the different paths in the code so that it runs on anyone's computer (once they've cloned the repo)