Create nat. gas cost, 6-19 to 6-21, no coops.csv
natural gas costs as given by PSC (see source files) for 25 months. This file is unfortunately not comprehensive of all of the natural gas providers in the state and we're missing at least one somewhat big one (either in the northwest or southeast - because the names in the file don't align to those in the PSC's state utility regions map it's not entirely clear which one).